Numpy float32

+++ XRD.py
@@ -128,7 +128,7 @@ class XRD(Dataset):
         data = np.load(self.data[root][idx], allow_pickle=True).item()
         x, y = data['features'], data['labels7' if self.num_classes == 7 else 'labels230']
 
-        return x, y
+        return x.astype(np.float32), int(y)
 
 
 # Verify or download data
